Reformatted File Locations and folders
the test folders should contain files that are not approved for the final game yet. The game folders should only contain files that are at or close to a finished state. For coders, I recommend creating a separate folder within the test folders for your own testing files in case you need to get someone else's test files. This will help avoid naming conflicts or losing track of the files you need.